improvement of control of ply2bin arguments
