+++ src/project.hpp
@@ -123,107 +123,5 @@ bool projectPointCloud (
            const double & scale,
            const std::vector < sensorData > & vec_sensorData );
 
-/*********************************************************************
-*  export projected point cloud to json file
-*
-**********************************************************************/
-
-/*! \brief Export projected point cloud in a json
-*
-* Given a list of 3D point and corresponding pixels on stiched EQR panorama,
-* export theses informations to a json file.
-*
-* \param poseName         Name of the pose file. JSON export will have the same
-* \param vec_sensorData   Calibration informations
-* \param scale            Scale factor to obtain metric point cloud
-* \param pointAndPixels   List of 3D points and associated EQR pixels
-*
-* \return  Nothing
-*/
-
-void  exportToJson (  const std::string  poseFile,
-                      const std::vector < sensorData > & vec_sensorData,
-                      const double  scale,
-                      std::vector < std::pair < std::vector <double>, std::vector <double > > > pointAndPixels
-);
-
-/*********************************************************************
-*  export point cloud to json file
-*
-**********************************************************************/
-
-/*! \brief Export point cloud in ply format into json file
-*
-* Given a list of 3D point, export it to a json file.
-*
-* \param jsonName         Name of json file in c-string array
-* \param pointAndPixels   List of 3D points and associated EQR pixels
-*
-* \return  Nothing
-*/
-
-void  pointCloudToJson ( const char * jsonName,
-                         std::vector < std::pair < std::vector <double>, std::vector <unsigned int> > > pointAndColor
-);
-
-
-/*********************************************************************
-*  load calibration data related to elphel cameras
-*
-**********************************************************************/
-
-/*! \brief Load calibration information
-*
-* Given a mac address and a mount point, load all information needed to do
-* the point cloud projection on the panorama.
-*
-* \param vec_sensorData   Array used to store the sensor calibration information.
-* \param sMountPoint      Camera folder mount point
-* \param smacAddress      Elphel camera's mac address
-*
-* \return  bool value telling if the loading was sucessfull or not.
-*/
-
-bool  loadCalibrationData(std::vector < sensorData > & vec_sensorData,
-                          const std::string & sMountPoint,
-                          const std::string & smacAddress);
-
-/*********************************************************************
-*  load point cloud
-*
-**********************************************************************/
-
-/*! \brief Ply reader
-*
-* Given a point cloud in ply format, read the file and store the corresponding
-* 3D points in a vector of pair < position, color >
-*
-* \param vec_sensorData   Array used to store the sensor calibration information.
-*
-* \return  bool value telling if the loading was sucessfull or not.
-*/
-
-bool loadPointCloud ( const char * fileName ,
-                      vector< std::pair < vector <double >, vector<unsigned int> > > & pointAndColor
-);
-
-/*********************************************************************
-*  load rig pose
-*
-**********************************************************************/
-
-/*! \brief Pose file parser
-*
-* Given a rig pose file generated by openMVG, read it and store the information
-* in an array of double.
-*
-* \param  fileName   rig pose filename
-* \param  rigPose    The loaded rig pose
-*
-* \return  bool value telling if the loading was sucessfull or not.
-*/
-
-bool  loadRigPose ( const std::string & fileName,
-                    vector< std::vector<double> > & rigPose );
 
 #endif /* PROJECT_HPP_ */
